% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clonoNames}
\alias{clonoNames}
\alias{clonoNames,clonoStats-method}
\alias{clonoNames,SingleCellExperiment-method}
\title{Get clonotype names}
\usage{
clonoNames(object)

\S4method{clonoNames}{clonoStats}(object)

\S4method{clonoNames}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{a \code{\link{clonoStats}} object or
\code{SingleCellExperiment} object containing \code{clonoStats} results.}
}
\value{
The names of the clonotypes summarized in the input \code{clonoStats}
object.
}
\description{
Get clonotype names
}
\examples{
data('contigs')
cs <- clonoStats(contigs)
clonoNames(cs)

}
