% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMquant.R
\name{EMquant}
\alias{EMquant}
\alias{EMquant,SplitDataFrameList-method}
\alias{EMquant,SingleCellExperiment-method}
\title{Clonotype quantification with the EM algorithm}
\usage{
EMquant(x, ...)

\S4method{EMquant}{SplitDataFrameList}(
  x,
  sample = "sample",
  type = NULL,
  method = c("python", "r"),
  thresh = 0.01,
  iter.max = 1000
)

\S4method{EMquant}{SingleCellExperiment}(x, TCRcol = "contigs", sample = NULL, ...)
}
\arguments{
\item{x}{A \code{SplitDataFrameList} object containing V(D)J contig
information, split by cell barcodes, as created by \code{readVDJcontigs}.
Alternatively, a \code{SingleCellExperiment} object with such a
\code{SplitDataFrameList} in the \code{colData}, as created by
\code{addVDJtoSCE}.}

\item{...}{additional arguments.}

\item{sample}{The name of the column in \code{x} (or the \code{colData} of
\code{x}, for \code{SingleCellExperiment} objects) that stores each cell's
sample of origin. Alternatively, a vector of length equal to \code{x} (or
\code{ncol(x)}) indicating the sample of origin. Providing this information
can dramatically speed up computation and avoid unwanted cross-talk between
samples.}

\item{type}{The type of VDJ data (\code{"TCR"} or \code{"BCR"}). If
\code{NULL}, whichever is most common in \code{x}.}

\item{method}{Which implementation to use. Options are \code{'r'} or
\code{'python'} (default).}

\item{thresh}{Numeric threshold for convergence of the EM algorithm.
Indicates the maximum allowable deviation in a count between updates.}

\item{iter.max}{Maximum number of iterations for the EM algorithm.}

\item{TCRcol}{The name of the column in the \code{colData} of \code{x} that
contains the VDJ contig data (only applies if \code{x} is a
\code{SingleCellExperiment}).}
}
\value{
Creates a sparse matrix (\code{dgRMatrix}) of cell-level clonotype
assignments (cells-by-clonotypes). If \code{x} is a
\code{SingleCellExperiment}, this matrix is added to the \code{colData}
under the name \code{clono}.
}
\description{
Assign each cell (with at least one V(D)J contig) to its most
likely clonotype with the EM algorithm. For ambiguous cells, this leads to
proportional (non-integer) assignment across multiple possible clonotypes.
}
\details{
This quantification method defines a clonotype as a pair of specific
chains (alpha and beta for T cells, heavy and light for B cells) and
attempts to assign each cell to its most likely clonotype. Unlike other
quantification methods, this can lead to non-integer counts for cells with
ambiguous information (ie. only an alpha chain, or two alphas and one beta
chain).

We highly recommend providing information on each cell's sample of
origin, as this can speed up computation and provide more accurate results.
Because the EM algorithm shares information across cells, splitting by
sample can improve accuracy by removing extraneous clonotypes from the set
of possibilities for a particular cell.
}
\examples{
data('contigs')
counts <- EMquant(contigs)

}
