% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{example_contigs}
\alias{example_contigs}
\alias{contigs}
\title{Example V(D)J contig data}
\format{
A \code{SplitDataFrameList} object of length 24.
}
\source{
An extremely small subset of the TCR data from Braun, et al. (2020).
}
\usage{
data("example_contigs")
}
\description{
This is a small dataset consisting of 24 cells from 2 samples,
meant to demonstrate how certain functions in the \code{TCRseq} package
work.
}
\examples{
data("example_contigs")
contigs

}
\keyword{datasets}
