% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueQuant.R
\name{uniqueQuant}
\alias{uniqueQuant}
\alias{uniqueQuant,SingleCellExperiment-method}
\alias{uniqueQuant,SplitDataFrameList-method}
\title{Clonotype quantification for uniquely assignable cells}
\usage{
uniqueQuant(x, ...)

\S4method{uniqueQuant}{SingleCellExperiment}(x, TCRcol = "contigs", sample = NULL, type = NULL)

\S4method{uniqueQuant}{SplitDataFrameList}(x, sample = "sample", type = NULL)
}
\arguments{
\item{x}{A \code{SplitDataFrameList} object containing V(D)J contig
information, split by cell barcodes, as created by \code{readVDJcontigs}.
Alternatively, a \code{SingleCellExperiment} object with such a
\code{SplitDataFrameList} in the \code{colData}, as created by
\code{addVDJtoSCE}.}

\item{...}{additional arguments.}

\item{TCRcol}{The name of the column in the \code{colData} of \code{x} that
contains the VDJ contig data (only applies if \code{x} is a
\code{SingleCellExperiment}).}

\item{sample}{The name of the column in \code{x} (or the \code{colData} of
\code{x}, for \code{SingleCellExperiment} objects) that stores each cell's
sample of origin. Alternatively, a vector of length equal to \code{x} (or
\code{ncol(x)}) indicating the sample of origin.}

\item{type}{The type of VDJ data (\code{"TCR"} or \code{"BCR"}). If
\code{NULL}, this is determined by the most prevalent \code{chain} types in
\code{x}.}
}
\value{
Creates a sparse matrix (\code{dgRMatrix}) of cell-level clonotype
assignments (cells-by-clonotypes). If \code{x} is a
\code{SingleCellExperiment}, this matrix is added to the \code{colData}
under the name \code{clono}.
}
\description{
Assign cells that can be uniquely mapped to a single clonotype.
This results in fewer cells being used compared to other quantification
methods, but all assignments are unambiguous.
}
\details{
This quantification method defines a clonotype as a pair of specific
chains (alpha and beta for T cells, heavy and light for B cells) and only
assigns cells that are a perfect match for their clonotype, leaving others
unassigned, even if they contain V(D)J information.
}
\examples{
data('contigs')
counts <- uniqueQuant(contigs)

}
