% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDiversity.R
\name{calculateDiversity}
\alias{calculateDiversity}
\alias{calculateDiversity,matrix-method}
\title{Sample diversity estimation}
\usage{
calculateDiversity(k, ...)

\S4method{calculateDiversity}{matrix}(
  k,
  methods = c("all", "shannon", "simpson", "invsimpson", "chao1", "chaobunge"),
  scale_factor = 100,
  ...
)
}
\arguments{
\item{k}{A matrix of abundance values where rows are features and columns are
samples.}

\item{methods}{A character vector specifying which diversity measures to use
(default = \code{'all'}, see Details).}

\item{scale_factor}{Numeric scaling factor for handling non-integer counts
(see Details).}
}
\value{
A matrix of diversity estimates for each sample. Note that the
\code{'chaobunge'} method also includes an estimate of the standard error.
}
\description{
This function uses various methods to estimate the clonotypic
diversity of samples based on a matrix of clonotype abundances (samples are
columns).
}
\details{
Available methods are Shannon entropy (\code{'shannon'}), Simpson
index (\code{'simpson'}), inverse Simpson index (\code{'invsimpson'}),
Chao1 richness (\code{'chao1'}), and Chao-Bunge richness
(\code{'chaobunge'}). A special value of \code{'all'} is also allowed,
which will run all methods listed above.

The \code{'chao1'} and \code{'chaobunge'} estimates assume all
abundances are integers. When this is not the case for the input matrix,
\code{k}, all values are multiplied by the \code{scaling_factor} and
rounded to the nearest integer. The resulting estimate is then divided by
\code{scaling_factor} to return to the original scale. The
\code{'shannon'}, \code{'simpson'}, and \code{'invsimpson'} methods work
with any input type.
}
