% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clonoNames}
\alias{clonoNames}
\alias{clonoNames,clonoStats-method}
\title{Get clonotype names}
\usage{
clonoNames(cs)

\S4method{clonoNames}{clonoStats}(cs)
}
\arguments{
\item{cs}{a \code{\link{clonoStats}} object.}
}
\value{
The names of the clonotypes summarized in the input \code{clonoStats}
object.
}
\description{
Get clonotype names
}
\examples{
data('contigs')
cs <- clonoStats(contigs)
clonoNames(cs)

}
