% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarizeClonotypes}
\alias{summarizeClonotypes}
\alias{summarizeClonotypes,Matrix-method}
\alias{summarizeClonotypes,SingleCellExperiment-method}
\alias{summarizeClonotypes,matrix-method}
\title{Get sample-level clonotype counts}
\usage{
summarizeClonotypes(x, by = "sample", ...)

\S4method{summarizeClonotypes}{Matrix}(x, by)

\S4method{summarizeClonotypes}{SingleCellExperiment}(x, by, clonoCol = "clono")

\S4method{summarizeClonotypes}{matrix}(x, by)
}
\arguments{
\item{x}{A (usually sparse) matrix of cell-level clonotype counts (cells are
rows and clonotypes are columns). Alternatively, a
\code{\link[SingleCellExperiment]{SingleCellExperiment}} with such a matrix
stored in the \code{colData}.}

\item{by}{A character vector or factor by which to summarize the clonotype
counts. If \code{x} is a \code{SingleCellExperiment} object, this can also
be a character, giving the name of the column from the \code{colData} to
use as this variable.}

\item{...}{additional arguments.}

\item{clonoCol}{A character providing the name of the cell-level clonotype
counts matrix in the \code{colData} of \code{x} (default = \code{'clono'}).
Only applies when \code{x} is a \code{SingleCellExperiment} object.}
}
\value{
A matrix clonotype counts where each row corresponds to a unique
value of \code{by} (if \code{by} denotes sample labels, this is a matrix of
sample-level clonotype counts).
}
\description{
Takes a matrix of cell-level clonotype counts and sums them
within groups (typically samples).
}
\examples{
example(addVDJtoSCE)
sce <- EMquant(sce, sample = "sample")
sampleLevelCounts <- summarizeClonotypes(sce, by = "sample")

}
